+++ editor/debugger/debug_adapter/debug_adapter_protocol.h
@@ -35,7 +35,7 @@
 #include "core/io/stream_peer_tcp.h"
 #include "core/io/tcp_server.h"
 #include "editor/debugger/debug_adapter/debug_adapter_types.h"
-#include "scene/debugger/scene_debugger.h"
+#include "scene/debugger/scene_debugger_object.h"
 
 #define DAP_MAX_BUFFER_SIZE 4194304 // 4MB
 #define DAP_MAX_CLIENTS 8
