+++ editor/run/game_view_plugin.h
@@ -34,7 +34,7 @@
 #include "editor/debugger/editor_debugger_plugin.h"
 #include "editor/editor_main_screen.h"
 #include "editor/plugins/editor_plugin.h"
-#include "scene/debugger/scene_debugger.h"
+#include "scene/debugger/runtime_node_select.h"
 #include "scene/gui/box_container.h"
 
 class EmbeddedProcessBase;
